// Client/src/app/dtos/ingredient.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export class Ingredient {
    caption: string;
    kkal100: number;
    protein100: number;
    fat100: number;
    carbon100: number;
    id: string;
    comment: string;
}